firmware: Convert deprecated strncpy to strscpy
commit bf63cfdc05f61cf5f954ed502e1e1ed7c2944f8f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Since strncpy is deprecated and it does not append null character at the
end, we are using strscpy, which will return no. of bytes it copied and it
will append null character at the end. And we are passing len-1 to strscpy
allocating last byte for null character.

The issue is also reported by sparse warning as "In 'strncpy' specified
bound 255 equals destination size."

Signed-off-by: Ankush Mehtre <ankush.mehtre@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>