LF-7815 remoteproc: imx_rproc: support starting Cortex-M33 from non-TCM address for i.MX93
commit 680aa11c7bdaddf6bbffd74bc0a94ef67593b69b from
https://github.com/nxp-imx/linux-imx.git

The current implementation not pass the M33 firmware loading address to
Arm Trusted Firmware(ATF), so that implies the M33 firmware is using
address 0xFFE0000. If the M33 firmware using a new address, it will not
boot. So parsing the elf file to get the address of section ".interrupts"
which is the first section of the elf file as bootaddr and pass to ATF.

Reviewed-by: Ye Li <ye.li@nxp.com>
Signed-off-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>