LF-7857-1: arm64: dts: imx8mm-evk: Modify the PHY node for new BB
commit 0ac704698066f9775a55b88573a89d8c30c9673f from
https://github.com/nxp-imx/linux-imx.git

The base board of i.MX8MM/MN EVK has been re-designed due to EOL
component for the ethernet PHY. And the new PHY is RTL8211. In
order to support both the new and old base board with the same
software, we add the properties of RTL8211 to the phy node. But
notice that, this is a workaround and not "upstreamable". We will
find a better method to the upstream in future.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>