LF-7857-2: arm64: dts: imx8mn-evk: Modify the PHY node for new BB
commit  2742802a19f054c208efeb0457d7eb6d0a6e5203 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

The base board of i.MX8MM/MN EVK has been re-designed due to EOL
component for the ethernet PHY. And the new PHY is RTL8211. In
order to support both the new and old base board with the same
software, we add the properties of RTL8211 to the phy node. But
notice that, this is a workaround and not "upstreamable". We will
find a better method to the upstream in future.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>