drivers: octeontx2-pcicons: cleanup console de-initialization
commit 9f1b89c2ef9c755d825f42269d21a72e379dc0c6 from
git@git.assembla.com:cavium/WindRiver.linux.git

Move the de-initialization code from 'remove' to 'shutdown' in
order to accommodate the driver being built-in to the kernel
(i.e., the normal case).  If the driver is 'built-in', then the
'remove' function is never called.  This is technically more
correct and allows for improved inter-operation with U-Boot
console driver.

Add 'flushing' wait to the pci console de-initialization in
order to allow all pending data to be read by an attached host.

Use a more-correct timer deletion API to prevent races.

Signed-off-by: Rick Farrington <rfarrington@marvell.com>
Change-Id: I98ea5d96b49b526dae8e704586e1e50c516ccf82
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/92902
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Wojciech Zmuda <wzmuda@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Signed-off-by: Beniamin Sandu <beniamin.sandu@windriver.com>