usb: host: xhci-plat: donot reset host controller on resume when USB as wakeup source
commit 59a7111525d00a0b95789b4e041155837bb578d4 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

When USB is set as a wakeup source, then after the resume all the save
state during suspend transition, should be restored.

If 'snps,xhci-reset-on-resume' quirk is set then, on the resume transition,
the state machine ignores the restore and start states, and resets the host
controller forcefully.

Since the host controller was not able to restore the state machine of the
connected USB devices and triggered reset on resume, as a result, the USB
devices got disconnected.

USB Disconnection error message:
usb 2-1: USB disconnect, device number 2

To resolve the above issue, disabled the "snps,xhci-reset-on-resume"
quirk bit when the USB is set as a wakeup source and enabled it when
the USB is not a wakeup source.

Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>