octeontx2-pf: qos send queues management
commit 27956029fbacf93897b58970e28834670e617a64 from
git@git.assembla.com:cavium/WindRiver.linux.git

The number of Send queues (SQs) are decided on the device probe
which is equal to the number of online cpus. These SQs are allocated
and deallocated in interface open and close calls respectively.
This patch supports initializing  and deinitializing Send queues
dynamically.

Change-Id: Ife6c299d44b927668336bddf81d4636015130534
Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Signed-off-by: Hariprasad Kelam <hkelam@marvell.com>
Signed-off-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
[RQ: Drop the parts which have been already applied.]
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>