Revert "net: axienet: Use NAPI for TX completion path"
commit 20d21cf2b11626c3f9116458d2bef71670f4ffaa from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

This reverts commit 9e2bc267e78068b512d4409b884662f425adb1ec.

This patch reverts the mainline NAPI changes which include TX and RX.
This is a temporary revert because the mainline NAPI implementation
not working with MCDMA. Local RX NAPI implementation is added
in subsequent patches. TX NAPI is a new feature and can be added
after further testing.

Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Acked-by: Harini Katakam <harini.katakam@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>