Revert "net: axienet: implement NAPI and GRO receive"
commit 1752ae4532bffde3c159a2c9aff2f1baac1480de from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

This reverts commit cc37610caaf8d13a6ecb8afd1fe2ebc2424ff622.

This patch reverts the mainline NAPI changes which include TX and RX.
This is a temporary revert because the mainline NAPI implementation
not working with MCDMA. Local RX NAPI implementation is added in
subsequent patches. TX NAPI is a new feature and can be added after
further testing.

Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Acked-by: Harini Katakam <harini.katakam@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>