net: axienet: Fix issues when DRE is not enabled in the h/w
commit 41e7e58db56c23dbab4d0660a099b0ebbeef9d0f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

If DRE (Data realignment engine) is not enabled in the DMA h/w,
SW has to take care of the alignment of the buffers.

Currently driver is not handling alignment of buffers properly,
Resulting weird behaviour when try to test the ethernet interface
with these kind of designs.

This patch fixes this issue by allocating a pool of tx buffers
In the driver when DRE is not enabled in the h/w.
When there is an unaligned skb comes it will use those
Allocated tx buffers when DRE is not enabled in the h/w.

Note: When DRE is not enabled in the h/w the tx
Side performance will be very less as there is
a manual copy in the hard_xmit for this case.
Tx Perf Numbers on ZynqMP:
Without DRE: 248 Mbits/sec.
With DRE:   932 Mbits/sec.

Signed-off-by: Appana Durga Kedareswara Rao <appana.durga.rao@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Acked-by: Harini Katakam <harini.katakam@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>