net: xilinx: Add MRMAC support
commit b5752ab73abeced08d9b2c2d815c4a6b8f6b1458 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Add MRMAC support to AXI Ethernet driver and MCDMA.
Add support for GT reset mechanism for MRMAC subsystem to work.
This subsystem contains a common PLL reset mask register and
individual GT and data path reset bits for each lane. The GT mask
and reset bits have to be asserted in tandem whereas the data path
resets can be controlled independently. The sequence and delays
are as recommended. This patch adds support for 4x10G and 4x25G with
the former tested.

Signed-off-by: Harini Katakam <harini.katakam@xilinx.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Signed-off-by: Sarath Babu Naidu Gaddam <sarath.babu.naidu.gaddam@xilinx.com>
Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Acked-by: Harini Katakam <harini.katakam@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>