LF-6879: arm64: dts:imx8ulp-evk: enable output pad high drive strength
commit 7593a5a29c6eb4205f0a6a259f1fd6bba2f963c0 from
https://github.com/nxp-imx/linux-imx.git

The hardware defect of 8ulp-evk-revb2 will cause a glitch on the falling
edge of the spi clock waveform. This glitch will cause an extra pulse on
the clock digital signal, which leads to 8ULP-EVK-RevB2 as master can't
communicate well with 8ULP-EVK board through lpspi bus.

By enabling output pad high drive strength, the glitch is eliminated and
it can run the correct communication with 8Mhz and above speed when using
master mode with 8ulp-evk-revb2.

Fixes: 0b316213f7c3 ("MLK-25555-3 arm64: dts: imx8ulp: enable lpspi5")
Signed-off-by: Carlos Song <carlos.song@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>