media: dw9807-vcm: Add regulator support to the driver
commit 980949f0ac79dc94a10ecacdfef7d2dc23b0c1d1 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

Uses the regulator notifier framework so that the current
focus position will be restored whenever any user of the
regulator powers it up. This means that should the VCM
and sensor share a common regulator then starting the sensor
will automatically restore the default position. If they
have independent regulators then it will behave be powered
up when the VCM subdev is opened.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>