PCI: controller: Add Marvell OcteonTx2 PCIe Endpoint driver
commit 2fad69f9d549cb2d242e17b2273f42d5edbc0adb from
git@git.assembla.com:cavium/WindRiver.linux.git

This patch adds a driver that aids in configuring the PEM when OcteonTx2
is in PCIe Endpoint mode. Most of the endpoint configuration is done in
firmware. The driver sets up the BAR4 space which is 4MB chunks in 16
indices. This region will be used as mailbox kind of communication
channel from host to endpoint.

Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>
Change-Id: I1399cd567327d02707d246bdde439958dcbe7c87
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>