clocksource: stm: Add a new compatible for timestamping
commit aa363c1ee04dd2bf904f0e52b35df29c1f2de849 from
https://github.com/nxp-auto-linux/linux

"nxp,s32cc-stm-ts" compatible will be used by the STM driver the same
way the other compatible; the only difference is the counting started
during the probing.

It's useful when the STM is used as a HW timestamp source for other
modules. E.g., CAN timestamping.

Issue: ALB-10872
Issue: ALB-9261
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>