pinctrl: pinctrl-zynqmp: Add support for Versal platform
commit 50ae383b1af4ac8b231cee5f9bce5eb81a5d90ae from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

As an experimental solution for timebeing which will be reverted when
firmware support is in place to get MIO ID information, store pin numbers
(MIO ID) in descriptors.

Currently Firmware support to get the MIO attribute information(MIO ID) is
missing, so for timebeing MIO ID information is hardcoded in the driver.
Once Firmware support to get the MIO attribute information is in place,
will remove the hardcoded ID information in the driver and query the
firmware to get the MIO ID information.

Signed-off-by: Sai Krishna Potthuri <sai.krishna.potthuri@amd.com>
Signed-off-by: Swati Agarwal <swati.agarwal@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>