spi: spi-zynqmp-gqspi: Fix erase failure when tx buswidth is 4
commit 770c2ffc56b5b98f8991672ad51aedc1ce721596 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Irrespective of the tx buswidth, for erase commands the flash device
expects to receive the address on single line. But when tx buswidth
is set to 4 and the driver issues erase command, it sends the address
on 4 lines. This causes the erase operation to fail.
Update the driver to send the address, for all erase commands, on
single line.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>