mtd: spi-nor: Fix spi_nor_read for multi die flash connected in dual parallel mode
commit 802d9f8702b88a770306542015328d128dcb6bbd from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

When multi die flash chip(N25Q512) are connected in dual parallel mode.
flashcp fails at 64MB offset(so the 32MB boundary of the single chip).

In spi_nor_read a multiply of 2 was missing in the remaining page length
calculation which resulted in data corruption.

This patch updated spi_nor_read and adds the multiply by 2 in the remaining
page length calculation to resolve the issue.

Signed-off-by: Leon Leijssen <leon.leijssen@topic.nl>
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>