arm: Fix the "invalid constant" error when assembling entry-common.S
The commit 82f18d6a09c0 ("arm: merge fixup for duplicate
TIF_NOTIFY_SIGNAL and other TIF_*") tried to fix the duplicate macro
warning on rt kernel. But after that change, the new values of the TIF_*
would cause the following assembler errors:
  arch/arm/kernel/entry-common.S: Assembler messages:
  arch/arm/kernel/entry-common.S:103: Error: invalid constant (1000e0) after fixup
  arch/arm/kernel/entry-common.S:286: Error: invalid constant (1000e0) after fixup

So adjust the value of TIF_* to make the assembler happy. After this
change, the TIF_* macros are the same as the ones in linux-6.2.y-rt.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>