usb: dwc3: gadget: fix IRQ hiber not found error message
commit 770f568f1a43fe127dc3337473c353f07aa42254 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

In USB gadget mode, we have the "hiber" interrupt-names DT property to
support the hibernation feature on the zynqMp and "USB-wakeup" for the
wakeup feature on the versal platform. On versal, the following error
message appear, due to the unavailability of the interrupt-names DT
property "hiber".

Message:
dwc3 fe200000.usb: error -ENXIO: IRQ hiber not found

In this case, we do not want platform_get_irq_byname() to print an error
on failure for versal, so switch to platform_get_irq_byname_optional()
instead which does not print an error.

Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>