mailbox: llce: Encapsulate cached frames
commit 0880e7616a37b8b24e5452af6dc3f7916839365b from
https://github.com/nxp-auto-linux/linux

The LLCE Logger uses cached frames for the cases when a new frame is
popped and doesn't belong to the channel being processed. A similar
mechanism is needed for the RX path when a FIFO is shared between
multiple BCANs. This is the preparatory work for that case.

Issue: ALB-9819
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>