clk: zynqmp: pll: Remove the limit
commit 9ea2a0096156b687061931f9016027996d54b559 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

The range is taken care in the zynqmp_pll_round_rate. Remove the rate range
in the zynqmp_clk_register_pll() to prevent the early truncation of the
frequencies and also allow multiple combinations of child and parent to get
more accurate rates. There are many changes since 5.15 specifically commit
c80ac50cbb37 ("clk: Always set the rate on clk_set_range_rate") that
changed the behavior of the clock framework. So the issue is not observed
in the earlier release.

Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>