arm_mpam: resctrl: Add support for mbm total
commit 0bb7e8fe6b9395a2b8ae4e4ce3ee95b0a619eeea from
git@git.assembla.com:cavium/WindRiver.linux.git

Currently, the driver enables "mbm_local" if mpam class that is picked
to be "L3" resource control, and supports the bandwidth counter enumerated
as part of "L3" resource control. But this will be a problem when the
bandwidth counter is hosted by DDR MSC, and is enumerated as part of
"MB" mpam class.

To support this configurtaion, lets enabled mbm_total if the mpam class
picked to be "MB" resource control. mbm_local is still used for the
bandwidth counter enumerated as part of "L3" resource control.

Signed-off-by: Amit Singh Tomar <amitsinght@marvell.com>
Change-Id: Ida7c0367465ba9f1343af3226f475b862d0fa214
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>