CDX: add the cdx bus driver
commit cf0d3b967f67e24e3b83e8c8e0ed68f607bad6c8 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Introduce AMD CDX bus, which provides a mechanism for scanning
and probing CDX devices. These devices are memory mapped on
system bus for Application Processors(APUs).

CDX devices can be changed dynamically in the Fabric and CDX
bus interacts with CDX controller to rescan the bus and
rediscover the devices.

Signed-off-by: Nipun Gupta <nipun.gupta@amd.com>
Link: https://lore.kernel.org/r/20230217132830.3140439-2-nipun.gupta@amd.com
State: waiting
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>