v4l: xilinx: dprx: Add check for status1 in irq_handler
commit 69aea021c9b5572b83b60ffc98b01df7dbc9c462 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

When an interrupt is generated, status for both IRQ cause registers are
recovered. If there is an interrupt in status1 but no interrupt in status,
the interrupts in status1 are never handled. This patch fixes by adding
check in the irq_hanlder, so the interrupts in status and status1 are
handled.

Signed-off-by: Rajesh Gugulothu <rajesh.gugulothu@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>