v4l: xilinx: dprx: Add check for status1 in irq_handler
commit d621866e7963f68de7db7ef86b04352039c0573e from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

When an interrupt is generated, status for both IRQ cause registers are
recovered. If there is an interrupt in status1 but no interrupt in status,
the interrupts in status1 are never handled. This patch fixes by adding
check in the irq_hanlder, so the interrupts in status and status1 are
handled.

Signed-off-by: Rajesh Gugulothu <rajesh.gugulothu@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>