LF-8071-1 arm64: dts: imx93: use a reserved mem-ranges to constrain ele-mu dma-range
commit 63a35ad282ca0f77c5c9b698a9eb2d9f048dd2f6 from
https://github.com/nxp-imx/linux-imx.git

sentinel cannot read/write DDR in the range
- 0xE0000000  0xFFFFFFFF

while genrating random numbers with hwrng device,
sometimes output address is allocated from the range
which is not accessible to sentinel.

so we are reserving 1MB of ddr memory
to make sure that the memory is allocated
from the reserved mem range.

Fixes: 45df0f582836 ("LF-7285: firmware: imx: ele: Add support for ELE HW Random Number Generator")
Signed-off-by: Gaurav Jain <gaurav.jain@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>