octeontx2-af: Update correct mask to filter IPv4 fragments
commit 9d09c5c4c8f74eb902302874fcb5f3944c21a0fd from
git@git.assembla.com:cavium/WindRiver.linux.git

During the initial design, the IPv4 ip_flag mask was set to 0xff.
Which results to filter only fragmets with (fragment_offset == 0).
As part of the fix, updated the mask to 0x20 to filter all the
fragmented packets irrespective of the fragment_offset value.

Fixes: bdaf35e52234 ("octeontx2-af: Add NPC support to filter packetbased on IP fragment")
Signed-off-by: Suman Ghosh <sumang@marvell.com>
Change-Id: I5c11fd62d2090ad777d776c4c0103039a01ef4c7
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/97382
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>