MLK-25315: dma: pxp: add monochrome mode support for PS engine
commit 5367f9ac345ba70870bdb951be3011b892b33bab from
https://github.com/nxp-imx/linux-imx.git

When input format of PS engine is monochrome, such as Y8 or Y4,
the low 16 bits of PS_VBUF are used as the U/V data in the data
path instead of sourcing U/V data from external buffers. In this
case it represents a fixed value for U/V data.

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>