firmware: xilinx: Update xlnx_get_crypto_dev_data
commit 921d1a7895dac557e3e7c39613bf469e22191e64 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Currently xlnx_get_crypto_dev_data is passing only the API ID for the
feature check call.
With the recent update in the feature check API, now the module ID
should also be passed along with the API ID as the argument. Hence
updating the feature check call in xlnx_get_crypto_dev_data as well.

Also if the secure feature is present then the API
version will be returned in case of ZynqMP. In case of Versal, success
will be returned if the feature is present.
Hence, updated the check for the return value of feature check call in
xlnx_get_crypto_dev_data.

Signed-off-by: Harsha Harsha <harsha.harsha@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>