spi: fsl-lpspi: Convert to platform remove callback returning void
commit d26aa1f95c5819bd7ac52729fb96b3b515c4d329 from
https://github.com/nxp-auto-linux/linux

[ Upstream commit edd49c898751838009940fe38a7ad4c1d23224df ]

The .remove() callback for a platform driver returns an int which makes
many driver authors wrongly assume it's possible to do error handling by
returning an error code. However the value returned is (mostly) ignored
and this typically results in resource leaks. To improve here there is a
quest to make the remove callback return void. In the first step of this
quest all drivers are converted to .remove_new() which already returns
void.

Trivially convert this driver from always returning zero in the remove
callback to the void returning variant.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Link: https://lore.kernel.org/r/20230303172041.2103336-27-u.kleine-koenig@pengutronix.de
Signed-off-by: Mark Brown <broonie@kernel.org>

Issue: ALB-10320
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>