spi: cadence-quadspi: Fix ospi resume failure
commit f6904e79577cb8175c6aad7694a85e0862a3289d from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Before going into reset if the controller was operating in DDR mode, then
the dtr flag is set. After coming out of a reset the controller is always
in SDR mode but as the dtr flag is set the driver assumes that the flash is
in DDR mode so, while sending data to the flash it makes sure that the data
is in even byte count and the address is aligned this results in failure
during driver resume with "spi-nor spi0.0: resume() failed" error message
Update the driver to reset the dtr flag while operating in SDR mode.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>