usb: xhci: drop and add the endpoint context in xhci_fixup_endpoint()
commit  36619da03ee2062d8af81f272392ce935e980e79 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

Setting both the Drop and Add bits on the input context prevents the
corruption of split transactions seen with the BCM2711 XHCI controller,
which is a dwc3 variant.

This is a downstream feature that allows usbhid to restrict polling
intervals on mice and keyboards, and was only tested on a VL805 which
didn't complain about the fact the endpoint got added twice.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>