s32cc: flashmap: Reduce DTB partition size
commit b273e9baa91d9794e1a9fcbb1f44c8dd68a3b1bf from
https://github.com/nxp-auto-linux/linux

Our Linux S32CC DTB platforms size are all < 100 KB.
Keeping a 1 MB DTB partition for our QSPI Flash (with
a total size of 64 MB) is not optimal.

Therefore, as part of this commit, reduce the DTB size to ~200 KB,
and use the rest of ~800 KB for our Linux Kernel image, which is
more prone to increase in size.

Issue: ALB-9920
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>