drivers: gpio: add QIXIS CPLD gpio controller
commit 3e85d046daccd9b3ae757ea00244e9a02feca1a1 from
https://github.com/nxp-imx/linux-imx.git

Add support for the GPIO controller found on some QIXIS CPLDs in
Layerscape boards. For the moment, we are adding support for the one
found on the LX2160ARDB board. This driver is using gpio-regmap.

A GPIO controller has a maximum of 8 lines (all found in the same
register). Even within the same controller, the GPIO lines could be only
output or only input, this is why the driver also implements the
.get_direction() callback.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>