xilinx-ai-engine: Remove l2_mask bitmap and use array for l2_mask
commit 3f1a4da8fc9d332b7e13d31f12d8dc3f09bb1a22 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

In current implementation of l2_bitmap in aie_interrupt(), there is a
bug where storing the l2_mask. l2_mask_64 is initialized in for loop but
not initialized. Since it is not initialized, use of l2_mask_64 is
undefined and its value if compiler dependent.

Bitmap is not needed here. Driver essentially uses the l2_bitmap to
store 32 bit values of l2_mask.

This patch removes the l2_bitmap and uses array u32 * for storing the
l2_mask of l2 controller.

Signed-off-by: Govindarajulu Varadarajan <govindarajulu.varadarajan@amd.com>
Reviewed-by: Hua Jiang <hua.jiang@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>