drivers: media: i2c: imx708: Fix WIDE_DYNAMIC_RANGE control with long exposure
commit f32d3672fe2349b0ed3f2989dc64ac44f78339cb from
https://github.com/raspberrypi/linux.git rpi-6.6.y

Setting V4L2_CID_WIDE_DYNAMIC_RANGE was causing the long exposure
shift count to be reset, which is incorrect if the user has already
changed the frame length to cause it to have a non-zero value.

Because it only updates control ranges and doesn't set any registers,
the control can also be applied when the sensor is not powered on.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>