net: dpaa2: add support for 25G interfaces
commit 37f1cf5b5eb192d71962c6545a65656a3901c9eb from
https://github.com/nxp-imx/linux-imx.git

Add support in the dpaa2-eth driver so that it's able to also treat the
PHY_INTERFACE_MODE_25GBASER interface type.
Add MC_25000FD in the mac_capabilities field and add support so that the
driver is able to change from the MC known interface types to the Linux
known ones.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>