phy: lynx-28g: configure more equalization params for 10GBASER
commit 81e734ff404007748cbf4d606e022078c3f6d042 from
https://github.com/nxp-imx/linux-imx.git

We discovered that not all the equalization parameters for a lane were
configured upon an interface change. Configure the extra 2 registers
with the appropriate values for 10GBASE-R.

This was revealed when we added support for the 10G <-> 25G dynamic
configuration of the SerDes lane. These two interface types have
different values for the RCCR0 and TTLCR0 registers, whereas the 10G and
1G share the same values.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>