octeontx2-af: fix to get different rq mask
commit 467a68e2554acd8ddf2118154ba9961a87aed9e3 from
git@git.assembla.com:cavium/WindRiver.linux.git

Currently we are getting same RQ mask id for different profile
as RQ mask id calculation is based on logical AND operation
between mask and set. We need to get different RQ mask for
different profile if there is a change in either mask or set
or both. This patch fixes the same.

Fixes: 8bdf0536aa82 ("octeontx2-af: add support for CPT second pass")

Signed-off-by: Rakesh Kudurumalla <rkudurumalla@marvell.com>
Change-Id: I14b03bdad713cdf3651638bf8e393364b08f8da9
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/101269
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Nithin Kumar Dabilpuram <ndabilpuram@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>