HACK: misc: Add dma-buf to physical address exporter
commit a6d074da1a95aa77f6d6315d7c454623ff578a7e from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This is driver allows user-space to attach a DMA-BUF and receive back its
CPU physical address. This is a temporary solution to allow CMEM like
functionality from allocated buffers. This is a hack and this will be
removed when proper solutions are implemented.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>