gpio: regmap: add the .get_direction() callback
commit 6a17e781bf3d34e2bdb580153a108d01012b0671 from
https://github.com/nxp-imx/linux-imx.git

There are GPIO controllers such as the one present in the LX2160ARDB
QIXIS CPLD which are single register fixed-direction. This cannot be
modeled using the gpio-regmap as-is since there is no way to
present the true direction of a GPIO line.

In order to make this use case possible, add a new callback to the
gpio_config structure - .get_direction() - which can be used by user
drivers to provide the fixed direction per GPIO line.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>