arm64: dts: imx8dxl-evk: configure multiple queues on eqos
commit 359471d63584e7005eb3f43a9aa2daaf55e4e6c1 from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git

Eqos ethernet support five queues on hardware, enable these queues and
configure the priority of each queue. Uses Strict Priority as scheduling
algorithms to ensure that the TSN function works.

The priority of each queue is a bitmask value that maps VLAN tag priority
to the queue. Since the hardware only supports five queues, this patch maps
priority 0-3 to queues one by one, and priority 4-7 to queue 4.

The total fifo size of 5 queues is 8192 bytes, if enable 5 queues with
store-and-forward mode, it's not enough for large packets, which would
trigger fifo overflow frequently. This patch set DMA to thresh mode to
enable all 5 queues.

Signed-off-by: Tao Yang <Tao.Yang1@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>