LF-8907-3: i2c: imx-lpi2c: fix debug timeout error log when push rx command
commit ad38a24c288bfc4137f4eec33ce85d8579fc0827 from
https://github.com/nxp-imx/linux-imx.git

When the master receives long data, it needs to push the rx command
multiple times. At the same time, the fifo size is limited and it
will take time to enter the schedule. The timeout is triggered when
there is still no TXFIFO available for rx cmd writing to after one
second.

Fix the timeout error log.

Signed-off-by: Carlos Song <carlos.song@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>