firmware: ti_sci: Use dt provided fw name and address to load at suspend time
commit 7246c79252d598079ea8f3a306a62fc1b3324bc9 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Use request_firmware_direct to load the fs stub LPM firmware to a
provided memory region. The filename for the firmware is provided
in the device tree as "firmware-name".

Use a pm_notifier for loading of the low power firmware during
PM_SUSPEND_PREPARE phase so that it can be loaded from the rootfs
before it is suspended. It is possible in the future for this
firmware to require reload, so add a check to indicate that the
firmware is currently loaded so it is only loaded once.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Vibhore Vardhan <vibhore@ti.com>
Signed-off-by: Georgi Vlaev <g-vlaev@ti.com>
Tested-by: Roger Quadros <rogerq@kernel.org>
Reviewed-by: Vibhore Vardhan <vibhore@ti.com>
Reviewed-by: Andrew Davis <afd@ti.com>
Link: https://serenity.dal.design.ti.com/lore/linux-patch-review/20230413121227.35114-6-d-gole@ti.com
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>