arm64: dts: ti: Introduce k3-am625-sk-lpmdemo
commit c51e6d0a3daa12c0d8ef6b67e72fda574ec044eb from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Re-Introduce the k3-am625-sk-lpmdemo dtb which is a configuration for the
AM625 SK EVM to enable checkout of the minimal set of required devices for
low power mode.
This will be extended to enable devices as low power support for them is
validated or added.

Co-developed-by: Vibhore Vardhan <vibhore@ti.com>
Signed-off-by: Vibhore Vardhan <vibhore@ti.com>
Signed-off-by: Dhruva Gole <d-gole@ti.com>
Reviewed-by: Vibhore Vardhan <vibhore@ti.com>
Reviewed-by: Andrew Davis <afd@ti.com>
Signed-off-by: Dhruva Gole <d-gole@ti.com>
Link: https://serenity.dal.design.ti.com/lore/linux-patch-review/20230413121227.35114-9-d-gole@ti.com
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>