can: dev: add transceiver capabilities to xilinx_can
commit 96b5647c498fe19965c09f0b12458b51b8d277f2 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Currently the xilinx_can driver does not support adding a phy like the
"ti,tcan1043" to its devicetree.

This code makes it possible to add such phy, so that the kernel makes
sure that the PHY is in operational state, when the link is set to an
"up" state.

Signed-off-by: Marcel Hellwig <git@cookiesoft.de>
Link: https://lore.kernel.org/r/20230417085204.179268-1-git@cookiesoft.de
[mkl: call phy_power_off() after pm_runtime_put()]
[mkl: remove error message for phy_power_on() failure]
[mkl: update kernel-doc for struct xcan_priv]
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>
State: upstream (d7588f02e8d8a795702cf06cbbb4303c48b02a61)
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>