LF-8839-4 ARM64: dts: imx8mp: disable super speed park mode
commit 97a5349d936b08cf301730b59e4e8855283f815c from
https://github.com/nxp-imx/linux-imx.git

iMX8MP has the same issue like iMX8MQ, so enable this quirk.

See Synopsys STAR number 9001415732:

Impacted Configurations:
 Host or DRD-host mode (DWC_USB3_MODE == 1 || DWC_USB3_MODE == 2)

Version(s) affected:
 3.30b and earlier versions

Defect Summary:
Park mode:
Park mode is designed to improve scheduling and servicing when there is
a single asynchronous EP active. In this case, the controller prefetch
data/TRBs to do 3x burst size worth of packets. When park mode is
disabled, there could be some delay between the bursts on the USB, if
there is a latency exists on the xHC DMA bus. The delay between the
bursts on the USB and the latency on the DMA bus are related. Park mode
is intended to tolerate these delays on the USB in the single asynchronous
EP servicing mode.

Issue:
In case there are multiple endpoints active, the scheduler should not
enter park mode. But, if the asynchronous endpoints go in and out of
flow control frequently, due to this issue, scheduler can enter park
mode, and start scheduling process, while actually there is more than
one active asynchronous EP. Per scheduler implementation it should
fetch TRBs into the cache, and the number of fetched TRBs should not
exceed the number of free space in the TRB cache for the endpoint, but
ensuring the wraparound does not occur in the circular buffer even if
index 0 of the TRB cache is free. Due to this issue, the xHC may
calculate the wrong number of TRB cache space available. It can lead
to fetching more TRBs than there is space available for. This causes
overwriting of the next cache area after the TRB cache area. This
causes the controller to have unexpected behavior. The above-described
issue observed for SS BI (Bus Instance) and not observed for HS/FSLS BIs.

For both SS/HS use cases where the device responds with NRDY/NAK
frequently during the microframe, it tends to bring down the performance.
Additional information is available in the programming guide - GUCTL1[16]
register description. Even it is mentioned for HS only, the performance
degrade could happen for SS also.

System Usage Scenario:
Because of this defect Host controller will not operate as normal: i.e.,
Host controller cannot respond or have an unexpected behavior.

Consequence(s):
Decrease in robustness of IP
When issue occurs, requires Host controller to reset

Recommendation:
 Implement the software workaround

Workaround(s):
Disable park-mode:
 GUCTL1.PARKMODE_DISABLE_SS = 1;

Reviewed-by: Xu Yang <xu.yang_2@nxp.com>
Signed-off-by: Li Jun <jun.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>