macsec: Don't rely solely on the dst MAC address to identify destination MACsec device
commit b75a36cfaf65eab3312df7fa22a660420937e267 from
https://github.com/nxp-auto-linux/linux

Offloading device drivers will mark offloaded MACsec SKBs with the
corresponding SCI in the skb_metadata_dst so the macsec rx handler will
know to which interface to divert those skbs, in case of a marked skb
and a mismatch on the dst MAC address, divert the skb to the macsec
net_device where the macsec rx_handler will be called to consider cases
where relying solely on the dst MAC address is insufficient.

One such instance is when using MACsec with a VLAN as an inner
header, where the packet structure is ETHERNET | SECTAG | VLAN.
In such a scenario, the dst MAC address in the ethernet header
will correspond to the VLAN MAC address, resulting in a mismatch.

Issue: ALB-10605

Signed-off-by: Emeel Hakim <ehakim@nvidia.com>
Reviewed-by: Subbaraya Sundeep <sbhatta@marvell.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 7661351a54ec9a6a20203f94fd459a9360049045)
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>