LF-8690-2: dmaengine: fsl-edma-v3: support edma src id check if using ch_mux
commit 14f92316be210b5b3afb236fdd2c2c37c4028793 from
https://github.com/nxp-imx/linux-imx.git

Support edma src id check in order to make sure ch_mux uniqueness.

The dma_spec->args[0] passes src id. The src id is equal to the chanid,
if the sdma controller channel allocation is fixed. This case doesn't
need to check src id because repeated src id will fail when it requests
channel.

The src id isn't related to chanid if the sdma controller channel
allocation is dynamic. So need check src id, and write the srcid to the
ch_mux register.

This patch checks the src id if it's using. It can avoid setting the same
srcid in different ch_mux.

Reviewed-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Joy Zou <joy.zou@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>