watchdog: xilinx_wwdt: Add Versal window watchdog support
commit 7ebf4d06d387eed2d8d9859087ea091b1b0efbaa from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Versal watchdog driver uses window watchdog mode. Window watchdog
timer(WWDT) contains closed(first) and open(second) window with
32 bit width. Write to the watchdog timer within predefined window
periods of time. This means a period that is not too soon and a
period that is not too late. The WWDT has to be restarted within
the open window time. If software tries to restart WWDT outside of
the open window time period, it generates a reset.

Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Reviewed-by: Guenter Roeck <linux@roeck-us.net>
Link: https://lore.kernel.org/r/20230420104231.2243079-4-srinivas.neeli@amd.com
State: waiting
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>