media: ti: j721e-csi2rx: Enable DMA draining
commit 8a9c8625744b2049c43216d1ff49060dfde7b643 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Once frames start dropping due to lack of buffers, some data gets stuck
in the DMA pipeline somewhere. So the first DMA transfer after frame
drops gives a partial frame. This is obviously not useful to the
application and will only serve to confuse it. Issue a DMA transaction
to drain that up so that the next frame comes out clean.

Similarly when stopping a stream, the DMA terminate does not currently
ensure a graceful cleanup of the pending data in the fabric. Due to this
pending packet affecting the DMA framing, other contexts that are
streaming will also be interrupted as the CSI_RX_IF expects previous
packets to be fully consumed before any new line, end of frame or start
of frame can be sent. Thus perform ti_csi2rx_drain_dma() to drain any
pending data, also update the log level for the drain DMA failure from
dev_warn to dev_dbg during stop, even when this warning was seen, the
next frame was proper this could be due to the fact that DMACNTX is
disabled asynchronously and the pending data is shorter.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
[Drain DMA during stream stop]
Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>