drm/tidss: Rename "hw_videoport" to "vp_idx"
commit 95e8791b6213966a0b7f310b65b20e0fafb9eb90 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The DSS outputs DPI video signals through its Video Ports (VPs), and
based on the SoC version, a DSS can have as many as 4 VPs.
"hw_videoport" was used to refer the index of the VP in question, but
that can be confused with DSS output ports (OLDI, DPI).

Rename "hw_videoport" to "vp_idx" to avoid the confusion.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Reviewed-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>