HACK: drm/tidss: Update the clock request value for OLDI videoports
commit 79c3278a33a8df5888eeb70632339d52af9f209c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The OLDI TX(es) require a serial clock which is 7 times the pixel clock
of the display panel. When the OLDI is enabled in DSS, the pixel clock
input of the corresponding videoport gets a divided-by 7 value of the
requested clock.
Hack update the set clock by requesting 7 times the value.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Reviewed-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>