arm64: dts: ti: k3-am625-sk: Add DSS ports, HDMI tx & peripherals
commit a6a620145bb4f22758a692c212b3a8481cf6b2cb from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add DT nodes for sil9022 HDMI transmitter (tx), HDMI connector and the
HDMI fixed master clock on the am625-sk board.

Additionally, add and connect output port for DSS (vp2), input and output
ports for the sil9022 HDMI tx and the input port for the HDMI connector.

The sil9022 HDMI tx is connected on the i2c1 bus. The HDMI connector on
the board is of type "a". The clock frequency of the master clock that
supports the HDMI tx is 12.288 MHz.

The dpi output signals from the vp2 of DSS are fed into the sil9022 HDMI
transmitter. These signals are converted into HDMI signals which are
further passed on to the HDMI connector on the board, on which display
can be connected via appropriate HDMI cables.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Tested-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>