phy: ti: gmii-sel: Add support for SGMII mode
commit d2d762638a0f5dc8969cbf1faf7185e3dcab2241 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add support to configure the CPSW MAC's PHY in SGMII mode if the SoC
supports it. The extra_modes member of the phy_gmii_sel_soc_data struct
corresponding to the SoC is used to determine whether or not the SoC
supports SGMII mode.

Link: https://lore.kernel.org/r/20230309063514.398705-2-s-vadapalli@ti.com/
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: Roger Quadros <rogerq@kernel.org>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>