remoteproc: k4-m4: Add a remoteproc driver for M4F subsystem
commit 4572e31ebde2789bcb3bfd41396a2a309df82096 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The AM62x and AM64x SoCs of the TI K3 family has a Cortex  M4F core in
the MCU domain. This core is typically used for safety applications in a
stand alone mode. However, some application (non safety related) may
want to use the M4F core as a generic remote processor with IPC to the
host processor. The M4F core has internal IRAM and DRAM memories and are
exposed to the system bus for code and data loading.

A remote processor driver is added to support this subsystem, including
being able to load and boot the M4F core. Loading includes to M4F
internal memories and to any predefined external code/data memory. The
carveouts for external contiguous memory is defined in the M4F device
node and should match with the external memory declarations in the M4F
image binary. The M4F subsystem has two resets. One reset is for the
entire subsystem i.e including the internal memories and ther other, a
local reset is only for the M4F processing core. When loading the image,
the driver first releases the subsystem reset, loads the firmware image
and then releases the local reset to let the M4F processing core run.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>
[Martyn Welch: Modified to use common code refactored from DSP driver]
Signed-off-by: Martyn Welch <martyn.welch@collabora.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>