v4l: vxd-dec: Create mmu programming helper library
commit d765f3e0f52ddbf80c1e08429b27d361c1b477cf from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The IMG D5520 has an MMU which needs to be programmed with all
memory which it needs access to. This includes input buffers,
output buffers and parameter buffers for each decode instance,
as well as common buffers for firmware, etc.

Functions are provided for creating MMU directories (each stream
will have it's own MMU context), retrieving the directory page,
and mapping/unmapping a buffer into the MMU for a specific MMU context.

Also helper(s) are provided for querying the capabilities of the MMU.

Signed-off-by: Buddy Liong <buddy.liong@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Sidraya <sidraya.bj@pathpartnertech.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>