vxe-vxd: decoder: Fix input buffer size
commit 5e65513c3be5ee5d464017bcb5c9d6222228b7f4 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Optimize decoder Input buffer size. The Input buffer was
allocated with size equal to raw frame size of YUV422 which is not supported
by the driver. The frame size is mainly dependent on the
resolution (width and height) and also the bit depth (8/10/12 etc).
As we only support NV12 format, width * height * 1.5 should be sufficient
for holding an encoded frame.

Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>