vxe-vxd: decoder: Error handling of fatal condition
commit f3f52724b296b940f33ad99ce4567df3fd5bd0fe from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Handling of the fatal error response from the IP, setting
the state of frame to discard and resetting the IP. The error
recovery function is executed in the error simulation condition,
when emergency flag is not set, only in a non fatal condition.
When the emergency flag is set, the IP reset is handled in the
schedule_work() api of Interrupt handling.

Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>