LF-8918-1 clk: imx: Fix iMX8ULP SPLL2 MULT range
commit d2527032e32ea9a84b4533caf7e77fa7ceb07880 from
https://github.com/nxp-imx/linux-imx.git

The SPLL2 on iMX8ULP is different with other frac PLLs, it can
support VCO from 650Mhz to 1Ghz. According to RM, the MULT is
using a range from 27 to 54, not some fixed values. If using
current PLL implementation, some clock rate can't be supported.

Fix the issue by adding new type for the SPLL2 and use MULT range
to replace MULT table

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Reviewed-by: Jacky Bai <ping.bai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>