octeontx2-pf: Allow ntuple rule to direct packet to VF with higher Rx queue than its PF
commit 486afa4357d6c3b7c636423fc885de7455ee14ca from
git@git.assembla.com:cavium/WindRiver.linux.git

It is possible to add a ntuple rule which would like to direct packet to
a VF whose number queue is greater/less than its PF's queue number. For
example a PF can have 2 Rx queues but a VF created on that PF can have
8 queues. As of today, ntuple rule will reject rule because it is checking
the requested queue number against PF's number of Rx queues. As a part
of this patch if the action of a ntuple rule is to move a packet to
a VF's queue then the check is removed. Also, a debug information is
printed to aware user that it is user's responsibility to cross check if
the requested queue number on that VF is a valid one.

Signed-off-by: Suman Ghosh <sumang@marvell.com>
Change-Id: If25826d59163f1cc4294c5632b2888e3ff31286c
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/102697
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>