dt-bindings: gpu: pvr: combine rogue devices
commit 7d9f803d9bc8385dab4396d11da9fbdbec833eee from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Clean this bindings file up and combine the rogue devices now that J*
hybrid APM is up and running.

Drop the power island pd on BXS/8XE. The power island is controlled
directly by the firmware through the use of an MMR on all of our
platforms. Registering this domain with the Linux power system
interferes with this behavior.

Now all rogue devices behave the same and require the same dt
attributes.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>