staging: xilinx-tsn: warn user when temac and ep mac address dont match
commit e82aa60f4bd06513e53c1533ef75edb8d337258d from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

According to the IP design, the MSB 44 bits of the mac address assigned to
all the ports of the TSN switch should be the same.
When there's a mismatch in the MAC addresses, the unicast frame filter
in the Xilinx TSN switch, cannot be setup correctly. As a result, features
like preemption (dependent on the MAC address of the port), might not work
as expected.
Adding a warning in case there's a mismatch.

Signed-off-by: Pranavi Somisetty <pranavi.somisetty@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>