coresight: etm4x: Configure ETM to trigger on panic
commit 7bb568afb34e7cff163ba00dadc61f41e2c35722 from
git@git.assembla.com:cavium/WindRiver.linux.git

Configure ETM to generate a trigger on Ext out [0] on panic.
This is achieved by configuring a single address comparator
with the address of "panic" and by configuring Ext out [0] to
trigger on the comparator event.

This trigger can be used by the sinks to stop capture on panic.

Change-Id: Iee82dad6e65535abb59850f52fb3b71c542898fa
Signed-off-by: Linu Cherian <lcherian@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/104422
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Tanmay Jagdale <tanmay@marvell.com>
Reviewed-by: Anil Kumar Reddy H <areddy3@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>