rpmsg: Make MAX_RPMSG_BUF_SIZE configurable from kconfig
commit 95883eb3d9118a030a633214ec7b38d725c8ae2a from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

The payload size for RPMsg send and receive path is very use case
specific. These payloads are restricted by the Virtio buffer size.
And this is configurable by MAX_RPMSG_BUF_SIZE macro.
The size of the buffer can be changed only by patching
MAX_RPMSG_BUF_SIZE.

To avoid the patching, RPMSG_VIRTIO_BUF_SIZE
is prompted from kconfig to configure maximum virtio buffer
size.

This patch is one of the rejected solution proposed on
openamp forum. More details can be found on
https://lists.openampproject.org/archives/list/openamp-rp@lists.openampproject.org/message/4BAIDFUZHHI7H6C5QFRBMMAM2ILF4WHD/

This patch is temporary and not upstream solution.

Signed-off-by: Abhijit Gangurde <abhijit.gangurde@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>