drm/bridge: cdns-dsi: Fix cdns_dsi_attach
commit ce7779c50da266efac6bfaa7c45484edbb53a31a from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

In cdns_dsi_attach function, when we find the output->bridge,
we are using dsi node itself and not the sn65dsi86 node in
of_drm_find_bridge. Due to this, the bridge is not added
and we get EINVAL. Fix it so that the bridge can be added
properly.

Fixes: e19233955d9e ("drm/bridge: Add Cadence DSI driver")
Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>