HACK: drm/panel-edp: Fix atomic check errors for panel
commit 76c9ac825a89bcd691224f8c48cd435f643035ec from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When we do the tidss_encoder_atomic_check, if the connector does
not have the bus_format, the check fails and we get EINVAL.
This bus_format object was removed from panel-edp driver when
it was split from panle-simple driver. So add the bus_format
back to the panel_desc structure and set it in connector
display_info.

Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>