crypto: octeontx2: fix opcode incase of SGv2
commit 66ee9c38e62319439c1caa14177ca76656bb0b05 from
git@git.assembla.com:cavium/WindRiver.linux.git

In SGv2, the result is being written to dptr if opcode
has inplace bit set, but driver looks for the result
in rptr, as SGv1 writes the result in rptr even inplace
bit is set in Scatter Gather mode.
This patch fixes this issue by not setting the inplace
bit for SG mode.

Signed-off-by: Srujana Challa <schalla@marvell.com>
Change-Id: I9658f6a6bdab78098962e33e3a71a1ff69afa308
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/104319
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Nithin Kumar Dabilpuram <ndabilpuram@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>