crypto: octeontx2: fix opcode incase of SGv2
commit 85d04c486fa33e981e77a560760c6c85830ee28e from
git@git.assembla.com:cavium/WindRiver.linux.git

In SGv2, the result is being written to dptr if opcode
has inplace bit set, but driver looks for the result
in rptr, as SGv1 writes the result in rptr even inplace
bit is set in Scatter Gather mode.
This patch fixes this issue by not setting the inplace
bit for SG mode.

Signed-off-by: Srujana Challa <schalla@marvell.com>
Change-Id: I9658f6a6bdab78098962e33e3a71a1ff69afa308
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/104321
Reviewed-by: Nithin Kumar Dabilpuram <ndabilpuram@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>