s32cc: nvmem: Refactor SIUL NVMEM to sync with SCMI NVMEM
commit 4b96f9d2536f43b5368f2655acbc4bb6dd521d28 from
https://github.com/nxp-auto-linux/linux

Added implementation in the SIUL NVMEM driver for getting
"soc_major" and "soc_minor" versions separately, and refactored
the `s32cc_nvmem_get_soc_revision` to make separate calls to NVMEM
in order to get both the major and minor versions. The data provided
by the NVMEM SIUL cells is not shifted anymore now, but rather
stored in the LSByte of the `u32` buffer returned.

Also, for the `s32cc_nvmem_get_soc_revision` and the
`s32cc_nvmem_get_pcie_dev_id` helpers, removed the manual passing
of the cell name, since the functions' names are explicit enough.

Updated the `s32cc.dtsi` and `pci-s32cc.c` file to match the
refactorings to the NVMEM SIUL driver and the NVMEM helpers.

Issue: ALB-9554
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>