arm64: dts: ti: k3-j7200-main: Switch MAIN R5F cluster to Split-mode
commit 2a162713d1f99b100b2afcbfe17db2a4bebedcfc from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

J7200 SoCs has a R5F cluster in each of the MAIN and MCU domains, and both
of these are configured for LockStep mode at the moment. Switch the MAIN
R5F cluster to Split mode by default to maximize the number of R5F cores.
The MCU R5F cluster continues to be in the preferred LockStep mode.

Note that this configuration is the default for remoteproc mode (kernel
driver boots the R5F processors). These will be overridden for early-booted
remoteprocs through the corresponding IPC-only support in the K3 R5F
remoteproc driver.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Apurva Nandan <a-nandan@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>