remoteproc/ti-k3: Add support for graceful shutdown
commit 652c6fa3ae566c07e2ef584160dce3f0c1a8129b from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This patch adds support for graceful shutdown of the TI remoteproc
driver on a stop command. On some TI SoCs, the remote processor
(ex: M4F on AM64/AM62 SoCs) needs to be in quiescent state, before a
reset is asserted. This patch introduces a mailbox handshake mechanism
between the MPU and the remote processor to allow the remote processor
to gracefully shutdown its peripherals and enter a quiescent state.

In addition, add elements for IPC handshake completion and PM notifiers
to k3-rproc structure.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>