PCI: j721e: Add support to build pci-j721e as a kernel module
commit 89f24ccfd3529c9fbc7b92c3bb47e37426bcada1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

pci-j721e driver can be build as a in-built kernel driver only, which is
not required as it is not used during boot time in most cases.
This change add support to build pci-j721e as a kernel loadable module.

J721e PCIe controller can work in both host mode and end-point mode.
In order to enable host mode driver and endpoint driver to be built
independently either as built-in or kernel module, the pcie-j721e.c driver
is refactored into following components:

1) pci-j721e-host.c: Driver used when PCIe controller has to be
initialized in host mode.

2) pci-j721e-ep.c: Driver used when PCIe controller has to be
initialized in endpoint mode.

3) pci-j721e.c: contains common code required in both modes.

Link: https://lore.kernel.org/all/20230513184620.2200916-3-a-verma1@ti.com/
Signed-off-by: Achal Verma <a-verma1@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>