spi: spi-omap2-mcspi: Use EOW interrupt for completion when DMA enabled
commit ea7d9188568669e6a1eb2e413b6c46bc208f0eb8 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

In MCSPI controller EOW interrupt is triggered when the channel has
transmitted the set number of bytes in MCSPI_XFERLEVEL[31-16] WCNT,
this can be used to signal the completion of a TX/RX when the internal
FIFO is enabled, when DMA is enabled the internal FIFO is always enabled.
Waiting for the DMA completion adds unpredictable delays due to the
non-realtime completion calculation mechanism.

Remove the dma_tx_completion and dma_rx_completion and rely on the
MCSPI controller EOW interrupt to signal transaction completion.
This fixes the real-time performance issues in master and slave mode
when DMA was enabled which resulted from the DMA completion calculation
delays.

Since the MCSPI driver now uses internal mechanism to identify a transfer
completion we disable the TX and RX DMA completion callback and remove
DMA_PREP_INTERRUPT.

Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>
Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>