LF-10269-2: mfd: Add support for MAX96752 deserializer with LVDS support
commit 7c433d1a97e56e03b328418b9ac8d320f8a8b4e6 from
https://github.com/nxp-imx/linux-imx.git

The MAX96752/MAX96754 deserializers convert a single or dual-link GMSL
serial input to single or dual OLDI. They also send and receive
side-channel and peripheral control data, enabling full-duplex,
single-wire transmission of video and bidirectional data.

This driver adds the core support with separate drivers adding
support for video or audio.

GPIO support is included in the core part as this is not a GPIO
controller in the Linux sense. It performs GPIO tunneling, translating
GPIO pin transitions from local side to the remote side and viceversa.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>