arm64: dts: ti: k3-am68-sk-base-board: Add support for IO expander
commit 06f5008e2d1e2b5f46e75ac2c852ab634e752fcc from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

AM68 SK features a dedicated IO expander for CSI2 that will be used
for sensor module resetting and power-down pins. The IO expander is
connected to the main_i2c1. Add support for the same.  The main_i2c1
node is disabled by default and shall be enabled in the camera overlays.

Also, CSI2RX instances which are routed to the Samtec MIPI connector on
the base board thus enable them as well.

Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>