arm64: dts: ti: k3-am68-sk: Add overlay for CSI2 interfaced with OV5640 camera
commit d0c4eec966e051ec740ed77d5d5119256cd919e1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The OV5640 camera sensor is interfaced with CSI2RX through a MIPI header.
The sensor shall be configured through the I2C which is interfaced with
the camera sensor via an I2C switch.

Add node for the I2C switch through which the camera sensor is connected
and add node for OV5640 camera connected to port 0 of the CSI bridge.

Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>