dt-bindings: clock: ehrpwm: Remove unneeded syscon compatible
commit ab0bba4a7292a6ac3aab792c54a5711d6556905d from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This node's register space is not accessed by any other node, which
is the traditional use for the "syscon" hint. It looks to have been
added here to make use of a Linux kernel helper syscon_node_to_regmap().
The Linux driver now uses a more appropriate helper that does not
require the hint, so let's remove it from the binding.

Signed-off-by: Andrew Davis <afd@ti.com>
Acked-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>