LF-9127-1: arm64: dtsi: imx8ulp: Support suspend/resume when camera is running
commit 334ef20c3215e7e9ae83571db114c9bf17ba3048 from
https://github.com/nxp-imx/linux-imx.git

The imx8ulp dtsi add flag for this function. The flag "runtime_suspend"
is to enable the runtime suspend when camera is running.

When system enter suspend mode, the capture pipeline SENSOR->CSI->ISI
can't work normally after resume, because the CSI/ISI power supplies are
lost. So if we want to resume normally, imx8ulp CSI and ISI should be
powered up, clock should be re-enabled, the CSI/ISI should be reset and
the registers should be re-configured. Meanwhile the pipeline should
follow the correct order.

Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: G.N. Zhou <guoniu.zhou@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>