wave5: decoder: Fix dmabuf import from downstream pool
commit 600f52ca6b17c9e70be3bb26f5cfeb3b752b6e5c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Due to hard limitaion of the IP of not able to program the address of
the v4l2 buffers dynamically, there is write to invalid buffers, which
produces green frames. This is resolved by registering the buffers used
at the init time and mapping with right index to address for valid
write of decoded data.

v2<-->v1: Fixed checkpatch errors and review comments.

Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Reviewed-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>