dt-bindings: ti-serdes-mux: Fix USB Type C swap define for J721S2 SoC
commit 826d456078f481aa5bc0e6f624f339ac6514ecce from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

If Lanes 1 and 3 are connected to the SerDes IP's USB PHY, Lanes 0
and 2 of the J721S2 SerDes WIZ are designated for USB type-C lane switch.

Only lane 2 USB swap was defined, and lane 0 USB swap was not.
Fix this by defining a USB swap macro for Lane 0.

Fixes: f5af1b824edc ("dt-bindings: ti-serdes-mux: Add USB Type C swap defines for J721S2 SoC")
Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>
Reviewed-by: Ravi Gunasekaran <r-gunasekaran@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>