usb: chipidea: imx: add one fsl picophy parameter tuning implementation
commit  8696d49659a9b9718648c9ed9bc13d1e27612c64 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

In some cases, the user may need to tune the rise/fall time of the
high-speed transmitter waveform for USB Certification. This will add
a parameter for this purpose. The value will be fetched from dtb and
finally written to the register.

Signed-off-by: Xu Yang <xu.yang_2@nxp.com>
Acked-by: Peter Chen <peter.chen@kernel.org>
Link: https://lore.kernel.org/r/20230627112126.1882666-3-xu.yang_2@nxp.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>