rpmsg: imx_rpmsg: Change the notifying to timeout mode
commit 46856a00ce52d8c3979b82a7358a268fcb4e51e0 from
https://github.com/nxp-imx/linux-imx.git

Currently, after the rpmsg link up it will use mailbox
channel in blocking mode to send the notifications. So
it will be stuck in the notification when the remote side
unable to respond for some reason, such as crashed.

To avoid interlock, this patch changes the notifying to
timeout mode, and treats the virtqueue as broken when
the remote side doesn't respond within 1 second.

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>