watchdog: dw_wdt: register notify callback to stop watchdog when shuting down system
On Intel SoC FPGA platform, Stratix10 or Agilex, the watchdog
is started in u-boot, so kernel will create a timer and work to
keep watchdog alive. But when user executes commands "halt",
"poweroff" or "shutdown", the system is reset when watchdog
triggers timeout. The root cause is that the watchdog not stopped
properly. So, register notify callback to stop watchdog when shuting
down system.

Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>