net: dpaa2-mac: skip pcs-handle search for phy-mode = "internal"
commit  a6d2791bdc558aa791f16236689b14fe9c9a7604 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

With DPAA2, internal PHYs implicitly mean the MTIP AN/LT blocks used for
copper backplanes. That driver is also in control of the PCS, and we are
only informed of the resolved link speed.

To avoid creating a phylink_pcs in that case, we make the check for
DPMAC_LINK_TYPE_BACKPLANE more complex, and exclude phy-mode = "internal"
from the prerequisite conditions for having a PCS under phylink's control.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>