arm64: dts: ti: k3-j721e: Remove PCIe endpoint nodes
commit cd7cb452d74130adfb2bef8002abd21a87caa38c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit a0cfd88d4a8a1106e9de5c3b03e68efe9e6249ec upstream.

These nodes are example nodes for the PCIe controller in "endpoint" mode.
By default the controller is in "root complex" mode and there is already a
DT node for the same.

Examples should go in the bindings or other documentation.

Remove this node.

Signed-off-by: Andrew Davis <afd@ti.com>
Link: https://lore.kernel.org/r/20230515172137.474626-2-afd@ti.com
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>