arm64: dts: ti: k3-j721e: Enable MDIO nodes at the board level
commit 54d89e2de8ef487f316c9aab94c058aff20f379a from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit b0efb45d126ee914bafca37a582ae6574560dd25 upstream.

MDIO nodes defined in the top-level J721e SoC dtsi files are incomplete
and will not be functional unless they are extended with a pinmux.

As the attached PHY is only known about at the board integration level,
these nodes should only be enabled when provided with this information.

Disable the MDIO nodes in the dtsi files and only enable the ones that
are actually pinned out on a given board.

Signed-off-by: Andrew Davis <afd@ti.com>
Link: https://lore.kernel.org/r/20230515172137.474626-5-afd@ti.com
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>