arm64: dts: ti: k3-j721e-beagleboneai64: Move camera gpio pinctrl to gpio node
commit 520bbd84c5092967cf9cc405abdf866d90c011d1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit d528c29fa7526adf7074fb166b99df7d5fd90670 upstream.

Move the GPIO mux configuration needed for camera module to work to the
GPIO node instead of the I2C node.

Camera nodes are maintained as overlay files, but the common mux is
always needed to ensure that camera probes fine and ensuring the mux
is configured as part of the GPIO module allows for the multiple
overlay files to be simpler.

Cc: Robert Nelson <robertcnelson@gmail.com>
Suggested-by: Udit Kumar <u-kumar1@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Link: https://lore.kernel.org/r/20230606182220.3661956-4-nm@ti.com
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>