crypto: atmel-aes: Add fallback for unsupported ESP header length
commit f3c4e2c8b0ec8742b5318bb7574132b9e8ad6173 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

Adding a software fallback for hmac(x)-cbc(aes) in cases where the ESP
header length is not a multiple of 4 words. The only ESP header length
that is not a multiple of 4 words is 6 words in this case, so this is the
only case tested.

As not all SoCs are affected by this limitation the HW version is used
to discriminate the SoCs. In this case HW versions 600 and 700 are
affected by this limitation.

To keep the hardware acceleration the check is put in place to use the
fallback for the highlighted case.

Signed-off-by: Ryan Wanner <Ryan.Wanner@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>