hwmon: (pwm-fan) Add fan speed register support
commit f2c5bc5afe1690c31aead4bca9069fc8b1c462e4 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

Some platforms include a fan-speed register that reports RPM directly
as an alternative to counting interrupts from the fan tachometer input.
Add support for reading a register at a given offset (rpm-offset) within
a block declared in another node (rpm-regmap). This indirection allows
the usual address mapping to be performed, and for address sharing with
another driver.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>