phy: lynx-28g: truly power the lanes up or down
commit 375b57c1a9dfddc346ba1d0e1e51ba0afcc63f35 from
https://github.com/nxp-imx/linux-imx.git

The current procedure for power_off() and power_on() is the same as the
one used for major lane reconfiguration, aka halting. But one can
observe that a halted lane does not cause, for example, the CDR of the
link partner to lose lock.

Implement the procedure mentioned in the block guide for powering down
a lane, and then back on.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>