media: ti: j721e-csi2rx: Make use of V4L2_CAP_IO_MC
commit c634ccb2086f43c26b2a0348ec507897d89ea412 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When V4L2_CAP_IO_MC flag is set (like it is by this driver) the
framework expects the video device to only enumerate the pixel formats
that it can convert to from the media-bus format supplied in the ioctl.

In the case of our CSI2RX driver, we cannot really convert to a
different pixelformat, so we should restrict enumeration with a 1-to-1
mapping between media bus and pixel formats.

Link: https://www.kernel.org/doc/html/latest/userspace-api/media/v4l/vidioc-enum-fmt.html#description
Fixes: 70f7e7ba6c66 ("media: ti: Add CSI2RX support for J721E")
Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>