drivers: iio: adc: versal-sysmon: Add I2C driver
commit f35a99dc1e0be36e2c4d1e99455b60c65b759276 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

This patch adds the I2C Driver for versal-sysmon.

This functionality requires that there is a functioning
I2C SysMon on the Versal chip (the PDI with a SysMon
with I2C must be loaded on the remote chip).

The device tree entry is added as an I2C child node.
Please see example node below:
i2c@5f {
	compatible = "xlnx,versal-sysmon";
	reg = <0x5f>;
	xlnx,numchannels = <7>;
	status = "okay";
	supply@0 {
		reg = <0>;
		xlnx,name = "vccaux";
	};
	supply@1 {
		reg = <1>;
		xlnx,name = "vccaux_pmc";
	};
	supply@2 {
		reg = <2>;
		xlnx,name = "vcc_pmc";
	};
	supply@3 {
		reg = <3>;
		xlnx,name = "vcc_psfp";
	};
	supply@4 {
		reg = <4>;
		xlnx,name = "vcc_pslp";
	};
	supply@5 {
		reg = <5>;
		xlnx,name = "vcc_soc";
	};
	supply@6 {
		reg = <6>;
		xlnx,name = "vp_vn";
	};
};

Signed-off-by: Conall O'Griofa <conall.ogriofa@amd.com>
State: pending
[michals: Squashed with
xilinx: Convert MODULE_LICENSE GPL v2 to GPL
]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>