phy: lynx-28g: replace LYNX_28G_SGMIIaCR1_SGPCS_DIS with 0
commit 17b731c840bb787065ff469ef1b0efeddbf44f03 from
https://github.com/nxp-imx/linux-imx.git

Now that lynx_28g_lane_rmw() no longer forces a macro name concatenation
inside its arguments, we no longer need to create macros with the "_DIS"
name in them which hold the value of 0. Sometimes there are actual
register fields which are named "DIS", and this becomes confusing,
because it isn't the case here.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>