s32cc: llce: Add support for LIN interrupt forwarding
commit ca2d6143ea0e7990e902fa2d546a5b320711a31c from
https://github.com/nxp-auto-linux/linux

In order to use Linflex modules from LLCE in UART mode, it is necessary
to completely disable the LIN Firmware of the LLCE and replace it with a
task that forwards the interrupts.

As stated in LLCE_firmware_user_guide, to enable the LIN interrupt
forwarding, the following steps should be done: - Set the
Lin_SharedMemory command ID (eCmdId) field to:
LLCE_LIN_CMD_ENABLEINTRFORWARD - Trigger an interrupt using CORE2CORE
module: LLCE_HOST2TX_ASSERT_INTERRUPT on the corresponding bit of the
command - Check the status in Lin_SharedMemory eReturnValue value

LIN interrupt forwarding is activated for all channels and can't be
deactivated until a reset of the core.

Issue: ALB-10318
Issue: ALB-10882
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>