s32cc: llce: Add Lin interrupt
commit ac854b3a3f374a4f28dcb2e06a05cd61bc828a8e from
https://github.com/nxp-auto-linux/linux

As stated in LLCE_firmware_user_guide, "the interface used between the
Host and the LLCE firmware is based on Core2Core module. There is only
one interrupt line that can be routed to a single Host. All the channels
shall be used on the same host. The Lin LLCE firmware is running on the
same internal core as the Transmission Protocol Processing Engine(TXPPE)
- CORE 2."

The same interrupt is used by all LLCE LIN channels.  As a concequence,
llce mailbox should work as an interrupt controller, forwaring the LLCE
interrupt from LLCE CORE TX (CORE 2) to all drivers registered to it.

Issue: ALB-10318
Signed-off-by: larisa.grigore <larisa.grigore@nxp.com>
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>