arm64: zynqmp: Rename xlnx,mio_bank to xlnx,mio-bank for DLC21
commit 4ecfc94f07fdfcc9a17860d0360425e08c578ebd from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

xlnx,mio_bank was used in past but it was renamed to xlnx,mio-bank because
'_' in property shoudln't be used. There is no impact on the platform
because if the properly is not defined bank 0 is default. Bank 0 and 1 have
the same configuration that's why there shouldn't be any issue.

Signed-off-by: Michal Simek <michal.simek@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>