phy: lynx-28g: refactor lane->interface to lane->mode
commit  44b0b351ec4e87ec0bd382ed74579d34adf7676f from
https://github.com/nxp-imx/linux-imx lf-6.1.y

The networking SerDes lanes can be configured in 2 ways, either as a
phy_interface_t when they are controlled by an Ethernet MAC driver, or
as an enum ethtool_link_mode_bit_indices when they are controlled by an
Ethernet PHY driver (for copper backplanes).

It would be desirable to add support for the latter, so create a new
lane operating mode namespace that can unify the 2 types above.

It is true that the phy_interface_t type also contains definitions for
PHY_INTERFACE_MODE_10GKR and PHY_INTERFACE_MODE_1000BASEKX, but those
were deemed to be mistakes, and shouldn't be used going forward, when
10GBase-KR and 1GBase-KX are really link modes.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>